+++ src/cmd/serve.rs
@@ -48,7 +48,7 @@ use ws::{Message, Sender, WebSocket};
 use errors::{anyhow, Context, Error, Result};
 use site::sass::compile_sass;
 use site::{Site, SITE_CONTENT};
-use utils::fs::{clean_site_output_folder, copy_file};
+use utils::fs::{clean_site_output_folder, copy_file, create_directory};
 
 use crate::fs_utils::{filter_events, ChangeKind, SimpleFileSystemEventKind};
 use crate::messages;
@@ -61,8 +61,8 @@ enum WatchMode {
     Condition(bool),
 }
 
-static METHOD_NOT_ALLOWED_TEXT: &[u8] = b"Method Not Allowed";
-static NOT_FOUND_TEXT: &[u8] = b"Not Found";
+const METHOD_NOT_ALLOWED_TEXT: &[u8] = b"Method Not Allowed";
+const NOT_FOUND_TEXT: &[u8] = b"Not Found";
 
 // This is dist/livereload.min.js from the LiveReload.js v3.2.4 release
 const LIVE_RELOAD: &str = include_str!("livereload.js");
@@ -430,6 +430,7 @@ pub fn serve(
     fast_rebuild: bool,
     no_port_append: bool,
     utc_offset: UtcOffset,
+    extra_watch_paths: Vec<String>,
 ) -> Result<()> {
     let start = Instant::now();
     let (mut site, bind_address, constructed_base_url) = create_new_site(
@@ -462,8 +463,8 @@ pub fn serve(
     // An array of (path, WatchMode, RecursiveMode) where the path is watched for changes,
     // the WatchMode value indicates whether this path must exist for zola serve to operate,
     // and the RecursiveMode value indicates whether to watch nested directories.
-    let watch_this = vec![
-        // The first entry is ultimtely to watch config.toml in a more robust manner on Linux when
+    let mut watch_this = vec![
+        // The first entry is ultimately to watch config.toml in a more robust manner on Linux when
         // the file changes by way of a caching strategy used by editors such as vim.
         // https://github.com/getzola/zola/issues/2266
         (root_dir_str, WatchMode::Required, RecursiveMode::NonRecursive),
@@ -473,6 +474,11 @@ pub fn serve(
         ("templates", WatchMode::Optional, RecursiveMode::Recursive),
         ("themes", WatchMode::Condition(site.config.theme.is_some()), RecursiveMode::Recursive),
     ];
+    watch_this.extend(
+        extra_watch_paths
+            .iter()
+            .map(|path| (path.as_str(), WatchMode::Required, RecursiveMode::Recursive)),
+    );
 
     // Setup watchers
     let (tx, rx) = channel();
@@ -502,6 +508,7 @@ pub fn serve(
     let ws_port = site.live_reload;
     let ws_address = format!("{}:{}", interface, ws_port.unwrap());
     let output_path = site.output_path.clone();
+    create_directory(&output_path)?;
 
     // static_root needs to be canonicalized because we do the same for the http server.
     let static_root = std::fs::canonicalize(&output_path).unwrap();
@@ -805,6 +812,23 @@ pub fn serve(
                             // No need to iterate over change group since we're rebuilding the site.
                             console::info("-> Config changed. The browser needs to be refreshed to make the changes visible.");
 
+                            if let Some(s) = recreate_site() {
+                                site = s;
+                            }
+                        }
+                        ChangeKind::ExtraPath => {
+                            let full_paths: Vec<&PathBuf> =
+                                change_group.iter().map(|(_, p, _)| p).collect();
+                            let combined_paths = full_paths
+                                .iter()
+                                .map(|p| p.display().to_string())
+                                .collect::<Vec<String>>()
+                                .join(", ");
+                            console::info(&format!(
+                                "-> {combined_paths} changed. Recreating whole site."
+                            ));
+
+                            // We can't know exactly what to update when a user provides the path.
                             if let Some(s) = recreate_site() {
                                 site = s;
                             }